requirements.txt
sentence-transformers
faiss-cpu
transformers
numpy
tqdm
